Incorporate the simpler llama code.
