+++ ihmc-high-level-behaviors/src/main/java/us/ihmc/behaviors/logic/ConditionNodeExecutor.java
@@ -29,6 +29,12 @@ public class ConditionNodeExecutor extends LeafNodeExecutor<ConditionNodeState, 
       }
    }
 
+   @Override
+   public void destroy()
+   {
+      llm.destroy();
+   }
+
    public CounterConditionExecutor getCounter()
    {
       return counter;
